+++ tools/LibraryVersionsGenerator/CSharpFileBuilder.cs
@@ -67,7 +67,7 @@ internal abstract class CSharpFileBuilder
 
     public abstract CSharpFileBuilder AddVersion(string version, string[] supportedFrameworks, string[] supportedPlatforms);
 
-    public abstract CSharpFileBuilder AddVersionWithDependencies(string version, Dictionary<string, string> dependencies, string[] supportedFrameworks, string[] supportedPlatforms);
+    public abstract CSharpFileBuilder AddVersionWithMetadata(string version, Dictionary<string, string> metadata, string[] supportedFrameworks, string[] supportedPlatforms);
 
     public abstract CSharpFileBuilder EndTestPackage();